Support SSL 2 and TLS 1.3 test dummies
The latest versions of openssl shipped in centos and most distros have
removed SSL 2 since it is deprecated for real-world use and very
unsafe. Further, they do not yet support TLS 1.3 which was added to
openssl in 1.1.1.

To provide test dummies for these protocol versions, this builds two
custom version of openssl and then uses openssl's s_server command as
the test server.

Unfortunately, SSL 2 is completely removed from OpenSSL 1.1.1 so we
still need to build 1.0.1.

This increases the time it takes to build the docker container
substantially.

Signed-off-by: Steven Danna <steve@chef.io>